Reorganize TransactionError, ParcelError into HistoryError, RuntimeError, SyntaxError
